# openapi.yaml
openapi: 3.0.3
info:
  title: Lombok Explorer API
  description: |
    API Contract untuk Lombok Explorer — Platform eksplorasi wisata, rekomendasi cerdas, 
    dan perencanaan perjalanan (itinerary planner) pulau Lombok, Nusa Tenggara Barat.
  version: 1.0.0
  contact:
    name: Lombok Explorer Team
    email: support@lombokexplorer.com
servers:
  - url: https://api.lombokexplorer.com/api/v1
    description: Production Server (v1)
  - url: http://localhost:8080/api/v1
    description: Local Development Server (v1)
  - url: http://localhost:8080/v1
    description: Legacy Compatibility Alias Server
paths:
...
  /auth/refresh-token:
    post:
      summary: Refresh expired access token using refresh token
      operationId: refreshToken
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /categories:
    get:
      summary: Get list of all travel categories
      operationId: getCategories
      tags:
        - Categories
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ==========================================
  # SEARCH
  # ==========================================
...
  /itineraries/generate:
    post:
      summary: Generate smart auto-itinerary based on days, preferences, and pace
      operationId: generateItinerary
      tags:
        - Itinerary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateItineraryRequest'
      responses:
        '200':
          description: Generated itinerary plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /itineraries:
    get:
      summary: Get user saved/created itineraries
      operationId: getItineraries
      tags:
        - Itinerary
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user itineraries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a custom itinerary
      operationId: createItinerary
      tags:
        - Itinerary
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryRequest'
      responses:
        '201':
          description: Itinerary created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /itineraries/{id}:
    get:
      summary: Get itinerary detail by ID
      operationId: getItineraryById
      tags:
        - Itinerary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Itinerary details with days and activities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDetailResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update existing itinerary
      operationId: updateItinerary
      tags:
        - Itinerary
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItineraryRequest'
      responses:
        '200':
          description: Itinerary updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete an itinerary
      operationId: deleteItinerary
      tags:
        - Itinerary
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Itinerary deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardActionResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /expenses:
    get:
      summary: Get user expense records
      operationId: getExpenses
      tags:
        - Expenses
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ExpenseDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create new expense item
      operationId: createExpense
      tags:
        - Expenses
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ExpenseDto'

  # ==========================================
  # JOURNALS
  # ==========================================
...
  /checklists:
    get:
      summary: Get user travel packing checklists
      operationId: getChecklists
      tags:
        - Checklists
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of packing checklists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChecklistDto'
    post:
      summary: Create custom checklist
      operationId: createChecklist
      tags:
        - Checklists
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  example: "Perlengkapan Snorkeling Gili"
      responses:
        '201':
          description: Checklist created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/ChecklistDto'

# ==========================================
# COMPONENTS & SCHEMAS
# ==========================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token for authorized requests
  responses:
    BadRequestError:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid query parameters"
            errorCode: "BAD_REQUEST"
    UnauthorizedError:
      description: Authentication is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Unauthorized access, please login"
            errorCode: "UNAUTHORIZED"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Destination not found"
            errorCode: "NOT_FOUND"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Internal server error occurred"
            errorCode: "INTERNAL_SERVER_ERROR"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - message
        - errorCode
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Resource not found"
        errorCode:
          type: string
          example: "DESTINATION_NOT_FOUND"
        details:
          type: array
          items:
            type: string
          nullable: true
    StandardActionResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"

    # --- ENUMS ---
    LombokRegion:
      type: string
      enum:
        - LOMBOK_SELATAN   # Kuta, Mandalika, Tanjung Aan, Merese, Mawun, Selong Belanak
        - LOMBOK_UTARA     # Senaru, Rinjani, Tiu Kelep, Bayan, Gili Islands
        - LOMBOK_BARAT     # Senggigi, Malimbu, Sekotong, Batu Layar
        - LOMBOK_TIMUR     # Sembalun, Pink Beach, Tetebatu
        - LOMBOK_TENGAH    # Sade, Sukarara, Praya
        - GILI_ISLANDS     # Trawangan, Meno, Air
      example: "LOMBOK_SELATAN"
...
    TravelStyle:
      type: string
      enum:
        - NATURE_ADVENTURE
        - BEACH_RELAXATION
        - CULTURE_HERITAGE
        - CULINARY_EXPLORER
        - PHOTOGRAPHY_SPOTS
        - FAMILY_FRIENDLY
      example: "BEACH_RELAXATION"
    BudgetLevel:
      type: string
      enum:
        - BUDGET
        - MID_RANGE
        - LUXURY
      example: "MID_RANGE"

    # --- AUTH & USER SCHEMAS ---
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
      example: "USER"
    UserDto:
      type: object
      required:
        - id
        - email
        - name
        - role
        - travelStyle
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: "usr_demo_lombok"
        email:
          type: string
          format: email
          example: "traveler@lombokexplorer.com"
        name:
          type: string
          example: "Wahyu Traveler"
        avatarUrl:
          type: string
          nullable: true
          example: "https://images.unsplash.com/photo-1535713875002-d1d0cf377fde?q=80&w=200"
        bio:
          type: string
          nullable: true
          example: "Travel enthusiast passionate about exploring tropical beaches & waterfalls."
        phone:
          type: string
          nullable: true
          example: "+6281234567890"
        role:
          $ref: '#/components/schemas/UserRole'
        travelStyle:
          $ref: '#/components/schemas/TravelStyle'
        preferences:
          type: array
          items:
            type: string
          example: ["Snorkeling", "Sunset Point", "Culinary"]
        createdAt:
          type: string
          format: date-time
          example: "2026-08-23T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-08-23T08:00:00.000Z"
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    AuthTokenData:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
        - user
      properties:
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresIn:
          type: integer
          example: 900
        tokenType:
          type: string
          example: "Bearer"
        user:
          $ref: '#/components/schemas/UserDto'
    AuthResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Authentication successful"
        data:
          $ref: '#/components/schemas/AuthTokenData'
...
    CategoryDto:
      type: object
      required:
        - id
        - name
        - slug
        - description
        - iconName
        - destinationCount
        - coverImageUrl
      properties:
        id:
          type: string
          example: "cat_beach"
        name:
          type: string
          example: "Pantai & Pesisir"
        slug:
          type: string
          example: "beach"
        description:
          type: string
          example: "Eksplorasi pantai pasir putih, teluk toska, dan ombak selancar kelas dunia."
        iconName:
          type: string
          example: "beach_access"
        destinationCount:
          type: integer
          example: 12
        coverImageUrl:
          type: string
          example: "https://images.unsplash.com/photo-1544644181-1484b3fdfc62?q=80&w=600"
    CategoryListResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Success fetching categories"
        data:
          type: array
          items:
            $ref: '#/components/schemas/CategoryDto'

    # --- RECOMMENDATION SCHEMAS ---
...
    ItineraryActivityDto:
      type: object
      required:
        - id
        - orderIndex
        - timeSlot
        - destinationId
        - destinationName
        - destinationCategory
        - imageUrl
        - estimatedDurationMinutes
        - activityNotes
        - estimatedCost
      properties:
        id:
          type: string
          example: "act_101"
        orderIndex:
          type: integer
          example: 1
        timeSlot:
          type: string
          example: "08:30 - 10:30"
        destinationId:
          type: string
          example: "dest_tanjung_aan"
        destinationName:
          type: string
          example: "Pantai Tanjung Aan"
        destinationCategory:
          type: string
          example: "Pantai & Pesisir"
        imageUrl:
          type: string
          example: "https://images.unsplash.com/photo-1544644181-1484b3fdfc62?q=80&w=600"
        estimatedDurationMinutes:
          type: integer
          example: 120
        activityNotes:
          type: string
          example: "Berenang di air tenang dan santai di warung kelapa muda."
        estimatedCost:
          type: number
          example: 25000
    ItineraryDayDto:
      type: object
      required:
        - dayNumber
        - title
        - date
        - activities
      properties:
        dayNumber:
          type: integer
          example: 1
        title:
          type: string
          example: "Hari 1: Mandalika Coastal Explorer"
        date:
          type: string
          example: "2026-09-10"
        notes:
          type: string
          example: "Fokus perjalanan di pesisir selatan Lombok Tengah."
        activities:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryActivityDto'
    ItineraryDto:
      type: object
      required:
        - id
        - title
        - description
        - coverImageUrl
        - totalDays
        - totalEstimatedBudget
        - travelStyle
        - budgetLevel
        - days
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          example: "itin_3days_lombok_classic"
        title:
          type: string
          example: "3 Hari Jelajah Pesona Lombok Selatan & Gili"
        description:
          type: string
          example: "Itinerary lengkap dari pantai eksotis Mandalika, bukit savana Merese, hingga snorkeling air kristal Gili Trawangan."
        coverImageUrl:
          type: string
          example: "https://images.unsplash.com/photo-1544644181-1484b3fdfc62?q=80&w=1000"
        totalDays:
          type: integer
          example: 3
        totalEstimatedBudget:
          type: number
          example: 1850000
        travelStyle:
          $ref: '#/components/schemas/TravelStyle'
        budgetLevel:
          $ref: '#/components/schemas/BudgetLevel'
        isSaved:
          type: boolean
          example: true
        days:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDayDto'
        createdAt:
          type: string
          format: date-time
          example: "2026-08-20T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-08-20T10:00:00Z"
    GenerateItineraryRequest:
      type: object
      required:
        - days
        - travelStyle
        - budgetLevel
        - preferredRegions
      properties:
        days:
          type: integer
          minimum: 1
          maximum: 7
          example: 3
        travelStyle:
          $ref: '#/components/schemas/TravelStyle'
        budgetLevel:
          $ref: '#/components/schemas/BudgetLevel'
        pace:
          type: string
          enum: [RELAXED, BALANCED, INTENSE]
          default: BALANCED
          example: "BALANCED"
        preferredRegions:
          type: array
          items:
            $ref: '#/components/schemas/LombokRegion'
          example: ["LOMBOK_SELATAN", "GILI_ISLANDS"]
        includeCulinary:
          type: boolean
          default: true
          example: true
    CreateItineraryRequest:
      type: object
      required:
        - title
        - totalDays
        - travelStyle
        - budgetLevel
        - days
      properties:
        title:
          type: string
          example: "Trip Liburan Keluarga Lombok"
        description:
          type: string
          example: "Rencana santai bersama keluarga."
        totalDays:
          type: integer
          example: 3
        travelStyle:
          $ref: '#/components/schemas/TravelStyle'
        budgetLevel:
          $ref: '#/components/schemas/BudgetLevel'
        days:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDayDto'
    UpdateItineraryRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        days:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDayDto'
    ItineraryListResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Success fetching itineraries"
        data:
          type: array
          items:
            $ref: '#/components/schemas/ItineraryDto'
    ItineraryDetailResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Success fetching itinerary"
        data:
          $ref: '#/components/schemas/ItineraryDto'

    # --- WEATHER SCHEMAS ---
...
    ExpenseDto:
      type: object
      required:
        - id
        - title
        - amount
        - currency
        - category
        - date
      properties:
        id:
          type: string
          example: "exp_001"
        title:
          type: string
          example: "Sewa Motor NMAX Mandalika (3 Hari)"
        amount:
          type: number
          example: 300000
        currency:
          type: string
          example: "IDR"
        category:
          type: string
          example: "TRANSPORT"
        date:
          type: string
          format: date-time
          example: "2026-08-23T08:00:00.000Z"
    CreateExpenseRequest:
      type: object
      required:
        - title
        - amount
        - category
      properties:
        title:
          type: string
          example: "Sewa Motor NMAX Mandalika"
        amount:
          type: number
          example: 300000
        currency:
          type: string
          default: "IDR"
          example: "IDR"
        category:
          type: string
          example: "TRANSPORT"
        notes:
          type: string
          example: "Sewa di bandara Lombok"
...
    ChecklistItemDto:
      type: object
      required:
        - id
        - item
        - isChecked
      properties:
        id:
          type: string
          example: "item_001"
        item:
          type: string
          example: "Kacamata Snorkeling & Masker"
        isChecked:
          type: boolean
          example: false
        quantity:
          type: integer
          example: 1
    ChecklistDto:
      type: object
      required:
        - id
        - title
        - items
      properties:
        id:
          type: string
          example: "chk_001"
        title:
          type: string
          example: "Perlengkapan Snorkeling Gili & Pantai"
        category:
          type: string
          example: "CLOTHING"
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItemDto'